Implemented basic testing
Implemented basic testing using Jest and React Native Testing Library.

Due to timebox expiration, I wasn't able to get tests for the navigation between the Pokemon List screen and Pokemon Details screen working. I believe this is due to an issue with finding the correct pressable view to force the navigation change in the testing library.

I also ran into issues correctly testing the info on the card. I ended up brute forcing the solution using a testId, but if time wasn't a concern I would debug this to find a way to test the info display in a way to better reflects what the user sees.

The loadMorePokemon function is also untested due to the expiration of the timebox I set on the project.